+++ authpass/lib/l10n/app_localizations_et.dart
@@ -365,6 +365,18 @@ class AppLocalizationsEt extends AppLocalizations {
   String get groupListFilterAppbarTitle => 'Filter by groups';
 
   @override
+  String get clearQuickUnlock => 'Clear Biometric Storage';
+
+  @override
+  String get clearQuickUnlockSubtitle => 'Remove saved master passwords';
+
+  @override
+  String get unlock => 'Unlock Files';
+
+  @override
+  String get clearQuickUnlockSuccess => 'Removed saved master passwords from biometric storage.';
+
+  @override
   String unexpectedError(String error) {
     return 'Unexpected Error: ${error}';
   }